feat(radiogo): RadioGo by infobit.cloud - complete radio streaming app
- Cyberpunk/matrix green theme (Orbitron + ShareTechMono, #00FF41)
- Radio-Browser.info API (50k+ stations, auto-updating links)
- just_audio + audio_service (background playback, lockscreen)
- GetX + Hive (favorites, recent, cache 24h TTL)
- Full IT+EN localization
- Radio Arcobaleno as default station
- Marquee scrolling, ICY metadata, search with debounce
- Station cards, mini player, full player, matrix rain splash

+++ lib/views/player/player_screen.dart
@@ -429,8 +429,8 @@ class PlayerScreen extends GetView<PlayerController> {
               color: AppColors.accentGreen.withAlpha(
                 (80 * glowIntensity).round().clamp(10, 80),
               ),
-              blurRadius: (20 * glowIntensity).round().clamp(4, 25),
-              spreadRadius: (4 * glowIntensity).round().clamp(0, 6),
+              blurRadius: (20 * glowIntensity).round().clamp(4, 25).toDouble(),
+              spreadRadius: (4 * glowIntensity).round().clamp(0, 6).toDouble(),
             ),
           ],
         ),
